% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{function to scan the landsld-df and check if there is a column called date}
\usage{
check_date(landsld)
}
\description{
function to scan the landsld-df and check if there is a column called date
}
