% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulated_pixel_value.R
\name{cumulate_pixel_value}
\alias{cumulate_pixel_value}
\title{Cumulate the soilmoisture-pixel value over time}
\usage{
cumulate_pixel_value(sm_image_paths = NULL)
}
\arguments{
\item{sm_image_paths}{A character vector containing the paths to soilmoisture images}
}
\description{
Cumulate the soilmoisture-pixel value over time
}
