% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sm_data.R
\name{get_sm_data}
\alias{get_sm_data}
\title{Function to extract the soilmoisture data for a set of landslides with a given date}
\usage{
get_sm_data(
  landsld = NULL,
  path_sm = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/soilmoisture/32632",
  days_before_window = 5,
  days_after_window = 0,
  point_buffer = NULL,
  aggre_fun = NULL,
  quiet = TRUE
)
}
\arguments{
\item{quiet}{Print an updating message on the status of the extraction}
}
\description{
This funtions takes a path to the folder of all the soilmoisture-tiffs and a geometry.
The geometry must have a column called \code{date}. This is necessasry in order to check if there is
soilmoisture data for the date of the landslide (plus, minus \code{days_before_window, days_after_window}).
It then extracts the soilmoisture values for the point, the buffered point, ot the polygon.
}
