% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_images_per_pixel.R
\name{how_many_images_per_pixel}
\alias{how_many_images_per_pixel}
\title{Mean Pixel value}
\usage{
how_many_images_per_pixel(sm_image_paths = NULL)
}
\arguments{
\item{sm_image_paths}{A character vector containing the paths to soilmoisture images}
}
\description{
Get the mean pixel value for a list of soilmoisture-image paths
}
