% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_extraction.R
\name{point_extraction}
\alias{point_extraction}
\title{function to extract the rainfall data for points or buffered points}
\usage{
point_extraction(landsld, matches, dates, point_buffer, aggre_fun)
}
\description{
function to extract the rainfall data for points or buffered points
}
