% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_extraction.R
\name{poly_extraction}
\alias{poly_extraction}
\title{function to extract the rainfall data for points or buffered points}
\usage{
poly_extraction(
  spatial.obj,
  paths_sm_tiffs,
  matches,
  tracks,
  swaths,
  point_buffer,
  times,
  aggre_fun
)
}
\arguments{
\item{matches}{A vector of class \code{Date} with the dates where the day of the landslide (plus minus the temporal buffer) intersect}

\item{aggre_fun}{A vector of one or multiple functions in order to aggregate the raster values that fall within the polygon.
For more see \code{\link{exact_extract}}}

\item{landsld}{One single polygon from a \code{sf}-object with many polygons (like on slope unit)}
}
\description{
function to extract the rainfall data for points or buffered points
}
